gemini_live_exhibition/requirements.txt
fastapi
uvicorn
google-generativeai
edge-tts
websockets
